Add classes for drag and drop handling
The DropManager class delegates drag and drop events to the currently
active DropWidget classes. A drop widget registers itself with the
DropManager when its parent property is changed.
On a drop event, it will check if the point is within the widget
and call its drop() function. Any class that wants to recieve drops
should implement this method and return True if it accepts the drop.
Otherwise, other widgets are tested.