Fix bug in DropWidget preventing correct drop
The DropWidget was listening to changes in its parent to determine if it
needs to add itself to or remove itself from the DropManager. This is
unreliable in the case that the widget is child of another widget, which
is added/removed form the windows widget tree. It now searches along its
path of parents for a root Node and adds/removes itself according to
that. This search is repeated every time one widget along the path
changes its parent.