Add card widget as a container for other widgets
A card widget displays a card similar to the material cards from google.
It uses an image (card.png) which consists of 9 regions from top-left to
bottom-right. The border of a card determines the sizes of the border.
Each child of the widget will be set to the position of the card widget
plus the border size and to a size equal to the size of the widget minus
two times the border.