Add status bar output to main app
The status bar takes messages through the message function. It will put
all messages into a list and display each message for duration seconds.
An exception are weak messages, which will be replaced as soon as a new
messgae is recieved (or after the timeout). Error messages get proiority
over normal messages. All messages will be automatically added to the
logfile.