Improve saving of links to files and folders
The default linktype will now check each path before saving. If the
target is a directory, it will add an explorer.exe command to the
vbs in order to open the directory. All other files will be opened
from a relative path by changing the working directory to the
directory where the file is located.