Restructure linktypes module
Instead of dynamically looking for linktype modules, the list of modules
is now fixed. The way the linktypeconfigs work has been changed.
The configs contain now the module and its name in addition to the settings.
The settings are now a dictionary instead of a list, so the keys are no
longer stored withing the settings classes. Serialization and deserialization
relies on the name of the linktype modules.